# docs/development/api/openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "description": "Standard error payload returned by every failed request.\n\nThe envelope is uniform across every error the server emits, including\nrouting fallbacks and extractor rejections (`API-039`).",
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/SelfLinks",
            "description": "Link to the request that produced the error."
          },
          "error": {
            "description": "Human-readable description of the error.",
            "example": "An error message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "_links"
        ],
        "type": "object"
      },
...
      "Link": {
        "description": "A single HAL link.",
        "properties": {
          "href": {
            "description": "Target URI reference of the link.",
            "type": "string"
          },
          "templated": {
            "description": "Whether `href` is a URI template, per RFC 6570.",
            "example": true,
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "href"
        ],
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "Payload to register a new user.\n\nThe requested `role` is granted only when the authenticated caller is\nallowed to create it: the Root Admin can register `Admin` or `Standard`\nusers, an `Admin` user can register `Standard` users only.",
        "properties": {
          "email": {
            "description": "Email address of the new user, when provided.",
            "format": "email",
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "description": "Password of the new user; never returned by the API.\n\nMust be at least 8 characters long and contain at least one symbol\n(a non-alphanumeric character).",
            "minLength": 8,
            "pattern": "[^A-Za-z0-9]",
            "type": "string",
            "writeOnly": true
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role to grant to the new user."
          },
          "username": {
            "description": "Username of the new user.",
            "maxLength": 100,
            "minLength": 4,
            "type": "string"
          }
        },
        "required": [
          "password",
          "role",
          "username"
        ],
        "type": "object"
      },
      "RegisterResponse": {
        "description": "User created by a successful registration.",
        "properties": {
          "email": {
            "description": "Email address of the newly registered user, when one was provided.",
            "format": "email",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/i64",
            "description": "Unique identifier of the newly registered user."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role granted to the newly registered user."
          },
          "username": {
            "description": "Username of the newly registered user.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "username"
        ],
        "type": "object"
      },
      "Role": {
        "description": "Role of a `User`.",
        "enum": [
          "ADMIN",
          "STANDARD"
        ],
        "type": "string"
      },
      "SelfLinks": {
        "description": "The `self` link of a HAL representation.\n\nEvery resource representation carries the URI it was served from, so clients\nnever construct or hardcode it themselves (`API-032`, `API-037`).",
        "properties": {
          "self": {
            "$ref": "#/components/schemas/Link",
            "description": "Link to the resource itself."
          }
        },
        "required": [
          "self"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API of the Mnemorium service",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Mnemorium API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/identity/register": {
      "post": {
        "description": "The Root Admin (user identifier `0`) may register `Admin` or `Standard`\nusers; an `Admin` user may register `Standard` users only; any other role is\nrejected with `403`.",
        "operationId": "post_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "User registered",
            "headers": {
              "Location": {
                "description": "URI of the newly registered user",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid payload"
          },
          "401": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Insufficient privileges to register the requested role"
          },
          "409": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Username or email is already taken"
          },
          "500": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Register a new user",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://0.0.0.0:4080/api/v1"
    }
  ],
...
}